pyproject.toml
[project]
name = "dungeon-service"
version = "0.1.0"
description = "ai powered dungeon master"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.18.1",
  "anthropic>=0.76.0",
  "fastapi>=0.128.0",
  "httpx>=0.28.1",
  "psycopg2-binary>=2.9.11",
  "pyjwt>=2.12.1",
  "python-dotenv>=1.2.1",
  "sqlmodel>=0.0.31",
  "typer>=0.21.1",
  "uvicorn>=0.40.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
extraPaths = ["."]

[tool.ruff.lint.per-file-ignores]
"src/models/*.py" = ["F821"]  # undefined-name — SQLAlchemy string references

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.14.0",
]
